+++ src/test/java/calculator/tokenizer/EmptyTokenizer.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.tokenizer;
 
 public class EmptyTokenizer implements Tokenizer {
