+++ src/test/java/calculator/tokenizer/NoneTokenizer.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.tokenizer;
 
 public class NoneTokenizer implements Tokenizer {
