+++ src/test/java/calculator/tokenizer/Tokenizer.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.tokenizer;
 
 public interface Tokenizer {
