+++ src/test/java/calculator/tokenizer/TokenizerFactory.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.tokenizer;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
